requirements.txt
flask>=2.3.0
gitpython>=3.1.0
python-dotenv>=1.0.0
